requirements.txt
# Core dependencies
telethon>=1.40.0
openai-whisper>=20231117